refactor: switch to unified template dialog
