--- src/components/dialogs/prompts/CustomizeTemplateDialog/index.tsx
@@ -1,57 +1,0 @@
-// src/components/dialogs/prompts/CustomizeTemplateDialog/index.tsx - Updated
-import React from 'react';
-import { getMessage } from '@/core/utils/i18n';
-import { useCustomizeTemplateDialog } from '@/hooks/dialogs/useCustomizeTemplateDialog';
-import { TemplateEditorDialog } from '../TemplateEditorDialog';
-
-export const CustomizeTemplateDialog: React.FC = () => {
-  const hook = useCustomizeTemplateDialog();
-
-  return (
-    <TemplateEditorDialog
-      // State
-      isOpen={hook.isOpen}
-      error={hook.error}
-      metadata={hook.metadata}
-      isProcessing={hook.isProcessing}
-      content={hook.content}
-      activeTab={hook.activeTab}
-      isSubmitting={hook.isSubmitting}
-      
-      // **NEW: Final content state**
-      finalPromptContent={hook.finalPromptContent}
-      hasUnsavedFinalChanges={hook.hasUnsavedFinalChanges}
-      modifiedBlocks={hook.modifiedBlocks}
-      
-      // Actions
-      setContent={hook.setContent}
-      setActiveTab={hook.setActiveTab}
-      handleComplete={hook.handleComplete}
-      handleClose={hook.handleClose}
-      
-      // **NEW: Final content actions**
-      setFinalPromptContent={hook.setFinalPromptContent}
-      applyFinalContentChanges={hook.applyFinalContentChanges}
-      discardFinalContentChanges={hook.discardFinalContentChanges}
-      updateBlockContent={hook.updateBlockContent}
-      
-      // Metadata update
-      setMetadata={hook.setMetadata}
-      
-      // UI state
-      expandedMetadata={hook.expandedMetadata}
-      toggleExpandedMetadata={hook.toggleExpandedMetadata}
-      activeSecondaryMetadata={hook.activeSecondaryMetadata}
-      metadataCollapsed={hook.metadataCollapsed}
-      setMetadataCollapsed={hook.setMetadataCollapsed}
-      secondaryMetadataCollapsed={hook.secondaryMetadataCollapsed}
-      setSecondaryMetadataCollapsed={hook.setSecondaryMetadataCollapsed}
-      customValues={hook.customValues}
-      
-      // Config
-      dialogTitle={getMessage('CustomizeTemplateDialog', undefined, 'Customize Template')}
-      dialogDescription={getMessage('CustomizeTemplateDialogDescription', undefined, 'Customize your prompt template')}
-      mode="customize"
-    />
-  );
-};
